添加测试空地一体的基础示例 (#3516)
* 解决安装hutb包后airsim包找不到的问题

* 添加测试空地一体的基础示例

* Remove residual virtual environment files to avoid "Permission denied" errors when creating virtual environments.

* roundabout: Python 3.12, 3.13, 3.14: No module named 'backports'

* 修复进行air示例测试时连不上测试端口的问题

+++ Util/download_from_git.py
@@ -440,7 +440,7 @@ if __name__ == "__main__":
             sys.exit(1)
         if os.path.exists(os.path.join(local_path, latest_file)):
             # 如果 hutb.zip 已经存在，则先删除
-            if os.path.exists(os.path.join(local_path, "hutb.zip")):
+            if os.path.exists(os.path.join(local_path, "hutb.zip"))  and latest_file != "hutb.zip":
                 print("Removing existing hutb.zip file: ", os.path.join(local_path, "hutb.zip"))
                 os.remove(os.path.join(local_path, "hutb.zip"))
             # 如果 latest_file 不是 hutb.zip，则重命名为 hutb.zip
